requirements.txt
azure-functions
azure-functions-worker
azure-eventhub
azure-eventgrid
azure-core
requests
